fix(evals): isolate analysis source attestation
